requirements.txt
binascii
pycryptodome
filemagic
binascii
nltk
